Improved Add and Delete Features

+++ crud/delete.php
@@ -14,25 +14,29 @@ $query = "SELECT * FROM pdf where f_id = $delkey";
 $result = mysqli_query($conn, $query);
 $row =  mysqli_fetch_assoc($result);
 
-// Delete the file from the 'pdf' folder
-$pdfFilePath = $row['file'];
-if (file_exists($pdfFilePath)) {
-    unlink($pdfFilePath);
-}
+if ($_SESSION['is_admin'] == '1' && $row['id'] == $profile) {
+    // Delete the file from the 'pdf' folder
+    $pdfFilePath = $row['file'];
+    if (file_exists($pdfFilePath)) {
+        unlink($pdfFilePath);
+    }
 
-// Delete the file from 'cover' folder
-$coverFilePath = $row['cover'];
-if (file_exists($coverFilePath)) {
-    unlink($coverFilePath);
-}
-$deleteQuery = "
+    // Delete the file from 'cover' folder
+    $coverFilePath = $row['cover'];
+    if (file_exists($coverFilePath)) {
+        unlink($coverFilePath);
+    }
+    $deleteQuery = "
     DELETE FROM resource_comment where r_id = $delkey;
     DELETE FROM resource_feedback where r_id= $delkey;
     DELETE FROM `pdf` WHERE f_id = $delkey;";
-$res = mysqli_multi_query($conn, $deleteQuery);
-if ($res) {
-    header('location:http://localhost/4thsemProj/pages/display.php');
-    exit;
+    $res = mysqli_multi_query($conn, $deleteQuery);
+    if ($res) {
+        header('location:http://localhost/4thsemProj/pages/display.php');
+        exit;
+    } else {
+        echo "Invalid request!";
+    }
 } else {
-    echo "Invalid request!";
+    header('location:http://localhost/4thsemProj/pages/display.php');
 }
